2 Main Changes
Cleaned up the entry of the product's price, so that any input is now rounded to 2 decimal places, and invalid characters are not permitted.

Also split the add_entry function in to 2. The first gathers the product details, and the second commits it to the database!

I think it's about there now, but if I've missed anything please do let me know.